phase2: relocate coeff_matrix types to circ.h
The coefficient-matrix carriers (enum stprep_kind,
struct data_coeff_block, struct data_coeff_matrix) describe the
algorithm-layer representation of the trial state, not the on-disk
schema, so their definitions belong with struct circ rather than
with the H5 I/O surface.  Move them to circ.h and leave the
data_state_prep_kind / data_coeff_matrix_load / _free declarations
in data.h alongside the rest of the loader API.

state_prep_coeff.h previously pulled data.h for these types; that
made circ.h -> state_prep_coeff.h -> data.h a candidate cycle once
the types reverse direction.  Forward-declare struct
data_coeff_matrix in state_prep_coeff.h instead (its only use is
as a pointer-to-const in two prototypes) and drop the data.h
include; data.h reciprocates with forward declarations of the
moved tags so data_state_prep_kind keeps its argument type
without including circ.h.
