+++ phase2/data.c
@@ -25,6 +25,7 @@
 #include "mpi.h"
 
 #include "log.h"
+#include "phase2/circ.h"
 #include "phase2/data.h"
 #include "phase2/world.h"
 #include <complex.h>
