+++ phase2/state_prep_coeff.c
@@ -19,6 +19,7 @@
 #include "combinations.h"
 #include "det_small.h"
 #include "log.h"
+#include "phase2/circ.h"
 #include "phase2/data.h"
 #include "phase2/qreg.h"
 #include "phase2/state_prep_coeff.h"
